+++ core-ng/src/main/java/core/framework/internal/async/VirtualThread.java
@@ -1,25 +1,7 @@
 package core.framework.internal.async;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import core.framework.internal.stat.Counter;
 
 public class VirtualThread {
-    public static final Stats STATS = new Stats();
-
-    public static class Stats {
-        final AtomicInteger count = new AtomicInteger(0);
-        final AtomicInteger maxCount = new AtomicInteger(0);
-
-        public void increase() {
-            int current = count.incrementAndGet();
-            maxCount.getAndAccumulate(current, Math::max);     // only increase active request triggers max active request process, doesn't need to handle when active requests decrease
-        }
-
-        public int maxCount() {
-            return maxCount.getAndSet(count.get());
-        }
-
-        public void decrease() {
-            count.decrementAndGet();
-        }
-    }
+    public static final Counter COUNT = new Counter();
 }