+++ core-ng/src/main/java/core/framework/internal/web/ShutdownHandler.java
@@ -1,27 +1,24 @@
 package core.framework.internal.web;
 
+import core.framework.internal.stat.Counter;
 import io.undertow.server.ExchangeCompletionListener;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.util.StatusCodes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.concurrent.atomic.AtomicInteger;
-
 /**
  * @author neo
  */
 public class ShutdownHandler implements ExchangeCompletionListener {
-    final AtomicInteger activeRequests = new AtomicInteger(0);
+    final Counter activeRequests = new Counter();
 
     private final Logger logger = LoggerFactory.getLogger(ShutdownHandler.class);
     private final Object lock = new Object();
-    private final AtomicInteger maxActiveRequests = new AtomicInteger(0);
     private volatile boolean shutdown;
 
     boolean handle(HttpServerExchange exchange) {
-        int current = activeRequests.incrementAndGet();
-        maxActiveRequests.getAndAccumulate(current, Math::max);     // only increase active request triggers max active request process, doesn't need to handle when active requests decrease
+        activeRequests.increase();
         exchange.addExchangeCompleteListener(this);
 
         if (shutdown) {
@@ -35,11 +32,6 @@ public class ShutdownHandler implements ExchangeCompletionListener {
         }
 
         return false;
-    }
-
-    // return max/peak active requests since last call, and reset max with current value
-    int maxActiveRequests() {
-        return maxActiveRequests.getAndSet(activeRequests.get());
     }
 
     void shutdown() {
@@ -63,7 +55,7 @@ public class ShutdownHandler implements ExchangeCompletionListener {
     @Override
     public void exchangeEvent(HttpServerExchange exchange, NextListener next) {
         try {
-            int count = activeRequests.decrementAndGet();
+            int count = activeRequests.decrease();
             if (count <= 0 && shutdown) {
                 synchronized (lock) {
                     lock.notifyAll();